refactor: consolidate skills test helpers into helpers_test.go
Move shared readXxxSkill and parseFrontmatter helpers into a single
helpers_test.go with generic readSkill(t, name) and parseFrontmatter
signatures, removing 6 near-identical per-file helper functions.

Closes #399
