+++ internal/skills/godark_create_phase_overview_test.go
@@ -8,15 +8,6 @@ import (
 	"github.com/phs/dark-factory/internal/skills"
 )
 
-func readCreatePhaseOverviewSkill(t *testing.T) string {
-	t.Helper()
-	data, err := fs.ReadFile(skills.SkillFiles, "godark-create-phase-overview/SKILL.md")
-	if err != nil {
-		t.Fatalf("reading godark-create-phase-overview/SKILL.md: %v", err)
-	}
-	return string(data)
-}
-
 func TestCreatePhaseOverviewSkillEmbedded(t *testing.T) {
 	_, err := fs.ReadFile(skills.SkillFiles, "godark-create-phase-overview/SKILL.md")
 	if err != nil {
@@ -25,7 +16,7 @@ func TestCreatePhaseOverviewSkillEmbedded(t *testing.T) {
 }
 
 func TestCreatePhaseOverviewSkillFrontmatterName(t *testing.T) {
-	content := readCreatePhaseOverviewSkill(t)
+	content := readSkill(t, "godark-create-phase-overview")
 	fm := parseFrontmatter(content)
 	if !strings.Contains(fm, "name: godark-create-phase-overview") {
 		t.Error("frontmatter missing 'name: godark-create-phase-overview'")
@@ -33,7 +24,7 @@ func TestCreatePhaseOverviewSkillFrontmatterName(t *testing.T) {
 }
 
 func TestCreatePhaseOverviewSkillFrontmatterDescription(t *testing.T) {
-	content := readCreatePhaseOverviewSkill(t)
+	content := readSkill(t, "godark-create-phase-overview")
 	fm := parseFrontmatter(content)
 	if !strings.Contains(fm, "description:") {
 		t.Error("frontmatter missing 'description' field")
@@ -41,7 +32,7 @@ func TestCreatePhaseOverviewSkillFrontmatterDescription(t *testing.T) {
 }
 
 func TestCreatePhaseOverviewSkillFrontmatterArgumentHint(t *testing.T) {
-	content := readCreatePhaseOverviewSkill(t)
+	content := readSkill(t, "godark-create-phase-overview")
 	fm := parseFrontmatter(content)
 	if !strings.Contains(fm, "argument-hint:") {
 		t.Error("frontmatter missing 'argument-hint' field")
@@ -49,7 +40,7 @@ func TestCreatePhaseOverviewSkillFrontmatterArgumentHint(t *testing.T) {
 }
 
 func TestCreatePhaseOverviewSkillFrontmatterDisableModelInvocation(t *testing.T) {
-	content := readCreatePhaseOverviewSkill(t)
+	content := readSkill(t, "godark-create-phase-overview")
 	fm := parseFrontmatter(content)
 	if !strings.Contains(fm, "disable-model-invocation: true") {
 		t.Error("frontmatter missing 'disable-model-invocation: true'")
@@ -57,35 +48,35 @@ func TestCreatePhaseOverviewSkillFrontmatterDisableModelInvocation(t *testing.T)
 }
 
 func TestCreatePhaseOverviewSkillReadsRoadmap(t *testing.T) {
-	content := readCreatePhaseOverviewSkill(t)
+	content := readSkill(t, "godark-create-phase-overview")
 	if !strings.Contains(content, "ROADMAP.md") {
 		t.Error("skill does not reference ROADMAP.md")
 	}
 }
 
 func TestCreatePhaseOverviewSkillReadsPlanningDoc(t *testing.T) {
-	content := readCreatePhaseOverviewSkill(t)
+	content := readSkill(t, "godark-create-phase-overview")
 	if !strings.Contains(content, "docs/planning/") {
 		t.Error("skill does not reference docs/planning/ for planning docs")
 	}
 }
 
 func TestCreatePhaseOverviewSkillExploresCodebase(t *testing.T) {
-	content := readCreatePhaseOverviewSkill(t)
+	content := readSkill(t, "godark-create-phase-overview")
 	if !strings.Contains(content, "docs/architecture.json") {
 		t.Error("skill does not instruct agent to use docs/architecture.json for codebase exploration")
 	}
 }
 
 func TestCreatePhaseOverviewSkillWritesToOverviewsDir(t *testing.T) {
-	content := readCreatePhaseOverviewSkill(t)
+	content := readSkill(t, "godark-create-phase-overview")
 	if !strings.Contains(content, "docs/phase-overviews/") {
 		t.Error("skill does not specify docs/phase-overviews/ as output directory")
 	}
 }
 
 func TestCreatePhaseOverviewSkillWarnsOnIncompletePhase(t *testing.T) {
-	content := readCreatePhaseOverviewSkill(t)
+	content := readSkill(t, "godark-create-phase-overview")
 	hasWarn := strings.Contains(content, "incomplete") || strings.Contains(content, "not marked")
 	if !hasWarn {
 		t.Error("skill does not warn when phase is not marked complete")
@@ -93,14 +84,14 @@ func TestCreatePhaseOverviewSkillWarnsOnIncompletePhase(t *testing.T) {
 }
 
 func TestCreatePhaseOverviewSkillRequiresRealWorldExamples(t *testing.T) {
-	content := readCreatePhaseOverviewSkill(t)
+	content := readSkill(t, "godark-create-phase-overview")
 	if !strings.Contains(content, "Example") {
 		t.Error("skill does not require real-world examples for each feature")
 	}
 }
 
 func TestCreatePhaseOverviewSkillGroundsInCodebase(t *testing.T) {
-	content := readCreatePhaseOverviewSkill(t)
+	content := readSkill(t, "godark-create-phase-overview")
 	hasGrounded := strings.Contains(content, "grounded") || strings.Contains(content, "actual codebase")
 	if !hasGrounded {
 		t.Error("skill does not emphasize grounding examples in actual codebase")
